rearrange some files
Move the checkbox code out into its own plugin file - no need to
complicate the ftplugin with this thing that should only be loaded once.

Put the syntax file in the after directory. The syntax here was getting
blown away by another markdown syntax file. We want the orgymode file to
be sourced on top of the existing syntax.